Add @firebase/firestore types
