+++ packages/firestore/src/platform_node/grpc_connection.ts
@@ -131,7 +131,7 @@ export class GrpcConnection implements Connection {
       log.debug(LOG_TAG, 'Creating Firestore stub.');
       const credentials = createHeaders(this.databaseInfo, token);
       this.cachedStub = {
-        stub: new this.firestore.Firestore(this.databaseInfo.host, credentials),
+        stub: new this.firestore.FirebaseFirestore(this.databaseInfo.host, credentials),
         token: token
       };
     }
