+++ packages/firestore/test/integration/api/batch_writes.test.ts
@@ -15,7 +15,7 @@
  */
 
 import { expect } from 'chai';
-import * as firestore from 'firestore';
+import * as firestore from '@firebase/firestore-types';
 
 import * as testHelpers from '../../util/helpers';
 import firebase from '../util/firebase_export';
