+++ packages/firestore/test/integration/api/type.test.ts
@@ -15,14 +15,14 @@
  */
 
 import { expect } from 'chai';
-import * as firestore from 'firestore';
+import * as firestore from '@firebase/firestore-types';
 import firebase from '../util/firebase_export';
 import { apiDescribe, withTestDb, withTestDoc } from '../util/helpers';
 
 import * as testHelpers from '../../util/helpers';
 
 apiDescribe('Firestore', persistence => {
-  function expectRoundtrip(db: firestore.Firestore, data: {}): Promise<void> {
+  function expectRoundtrip(db: firestore.FirebaseFirestore, data: {}): Promise<void> {
     const doc = db.collection('rooms').doc();
     return doc
       .set(data)
